fix: username unique check "23505" -> "NUNIQ"
I believe that the error code "23505" is being rewritten as "NUNIQ" and so this form is checking for the wrong error code. 

The error codes are stated in the [error_codes.md](/docs/error_codes.md) file, replaced using the [handleErrors.ts](@app/server/src/utils/handleErrors.ts) file and I am seeing "NUNIQ" being returned when using the starter.